Replaces all legacy services with the new framework.
